pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dynafield"
version = "0.0.9"
description = "Dynamic fields build up schema"
readme = "README.md"
authors = [
    { name = "Badr Elfarri", email = "badr@revvue.ai" },
    { name = "Magnus Mariero", email = "magnus@revvue.ai" }
]
maintainers = [
    { name = "Badr Elfarri", email = "badr@revvue.ai" },
    { name = "Magnus Mariero", email = "magnus@revvue.ai" }
]
license = "Apache-2.0"
requires-python = ">=3.12,<3.13"
keywords = ["dynamic fields", "toolkit", "async"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "uvicorn", # ASGI server for FastAPI (production/development)
    "pydantic[email]",
    "httpx",
    "fastapi",
    "uuid-utils",
    "strawberry-graphql[debug-server,fastapi]",
    "faststream[rabbit,otel]",
    "pika>=1.3.2",
    "rediskit>=0.0.27",
    "pyjwt>=2.10.1",
    "dotenv>=0.9.9",
    "opentelemetry-api>=1.39.0",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "polars>=1.35.2",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "alembic>=1.17.2",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "mypy",  # Mypy strawberry not supported in 1.18
    "pytest-asyncio",
    "types-zstd",
    "ariadne-codegen",
]
docker = [
    "dockerbuildsystem>=1.1.64",
]

[project.urls]
Homepage = "https://github.com/revvue-ai/dynafield"
Documentation = "https://github.com/revvue-ai/dynafield#readme"
Repository = "https://github.com/revvue-ai/dynafield"
"Bug Tracker" = "https://github.com/revvue-ai/dynafield/issues"
Changelog = "https://github.com/revvue-ai/dynafield/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["dynafield"]

[tool.setuptools.package-data]
dynafield = ["py.typed"]

[tool.isort]
line_length = 999
multi_line_output = 5

[tool.ruff]
line-length = 160
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"config.py" = ["E501"]
"app.py" = ["E402"]
"env.py" = ["E402"]
"migrate.py" = ["E722"]
"test_multi_document_agent.py" = ["E402", "F811"]

[tool.mypy]
strict = true
allow_untyped_decorators = true
ignore_missing_imports = true
plugins = [
    "pydantic.mypy",
    "strawberry.ext.mypy_plugin",
]

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "example.*",
    "build.*",
]
ignore_missing_imports = true
ignore_errors = true

#--- Ariadne codgen of revvue client code config -----------------------------------------------------------------------
[tool.ariadne-codegen]
remote_schema_url = "http://localhost:1000/graphql"
queries_path = "example"
target_package_name = "client"
target_package_path = "example"

[tool.ariadne-codegen.scalars.DateTime]
type = "datetime.datetime"
# ----------------------------------------------------------------------------------------------------------------------

